lib: Add error module
Separate the firmware errors with the firmware types.
